+++ src/main/java/fr/openmc/core/features/dream/registries/items/armors/creaking/OldCreakingLeggings.java
@@ -12,7 +12,7 @@ public class OldCreakingLeggings extends DreamEquipableItem {
     }
 
     @Override
-    public int getAdditionalMaxTime() {
+    public long getAdditionalMaxTime() {
         return 5;
     }
